refactor(api): split validateTokenAndGetType from user service
