+++ packages/api/src/modules/user/user.service.test.ts
@@ -432,76 +432,6 @@ describe("User Service", () => {
         });
     });
 
-    describe("validateTokenAndGetType", () => {
-        const FAKE_TOKEN = "FAKE";
-        const validUserReset = {
-            // userId: "1FR13J414N",
-            createdAt: new Date(),
-        } as unknown as WithId<UserReset>;
-        const user = {
-            profileToComplete: true,
-        } as unknown as UserDto;
-
-        beforeAll(() => {
-            mockedUserResetRepository.findByToken.mockResolvedValue(validUserReset);
-            mockedUserActivationService.validateResetToken.mockImplementation(() => ({ valid: true }));
-            mockGetUserById.mockResolvedValue(user);
-        });
-
-        afterAll(() => {
-            mockedUserResetRepository.findByToken.mockReset();
-            mockedUserActivationService.validateResetToken.mockReset();
-            mockGetUserById.mockReset();
-        });
-
-        it("should call find by token", async () => {
-            await userService.validateTokenAndGetType(FAKE_TOKEN);
-
-            expect(mockedUserResetRepository.findByToken).toBeCalledWith(FAKE_TOKEN);
-        });
-
-        it("should return true", async () => {
-            const actual = await userService.validateTokenAndGetType(FAKE_TOKEN);
-            expect(actual.valid).toBeTruthy();
-        });
-
-        it("should call validateResetToken", async () => {
-            await userService.validateTokenAndGetType(FAKE_TOKEN);
-
-            expect(mockedUserActivationService.validateResetToken).toHaveBeenCalledWith(validUserReset);
-        });
-
-        it("should return type is SIGNUP", async () => {
-            //@ts-expect-error: mock
-            mockedUserResetRepository.findByToken.mockResolvedValueOnce({
-                createdAt: new Date(),
-            });
-            const actual = (await userService.validateTokenAndGetType(
-                FAKE_TOKEN,
-            )) as TokenValidationDtoPositiveResponse;
-
-            expect(actual.type).toBe(TokenValidationType.SIGNUP);
-        });
-
-        it("should return type is FORGET_PASSWORD", async () => {
-            // @ts-expect-error: mock
-            mockedUserResetRepository.findByToken.mockResolvedValueOnce({
-                createdAt: new Date(),
-            });
-
-            // @ts-expect-error: mock
-            mockGetUserById.mockResolvedValueOnce({
-                profileToComplete: false,
-            });
-
-            const actual = (await userService.validateTokenAndGetType(
-                FAKE_TOKEN,
-            )) as TokenValidationDtoPositiveResponse;
-
-            expect(actual.type).toBe(TokenValidationType.FORGET_PASSWORD);
-        });
-    });
-
     describe("getUserWithoutSecret", () => {
         const EMAIL = "user@mail.fr";
